Change send/recv files hierarchy
binary_serialization.h -> send.h
binary_deserialization.h - recv.h

array-related send and recv implementation have been moved to
detail/array.h

The idea is to keep io-related stuff of a data type in single place.

send.h and recv.h contains common prototypes and functionality.
Data types io-related code has to be kept in separate files.
It looks more user friendly and models user code behavoiur -
io customization in separate file.
