+++ include/ozo/impl/async_request.h
@@ -4,7 +4,8 @@
 #include <ozo/detail/post_handler.h>
 #include <ozo/detail/timeout_handler.h>
 #include <ozo/impl/io.h>
-#include <ozo/binary_deserialization.h>
+#include <ozo/detail/array.h>
+#include <ozo/recv.h>
 #include <ozo/binary_query.h>
 #include <ozo/connection.h>
 #include <ozo/query_builder.h>
